ENH: do not use unsigned long but instead size_t for return type of filesize

+++ Source/Common/gdcmSystem.cxx
@@ -295,7 +295,7 @@ bool System::RemoveFile(const char* source)
 }
 
 // return size of file; also returns zero if no file exists
-unsigned long System::FileSize(const char* filename)
+size_t System::FileSize(const char* filename)
 {
   struct stat fs;
   if (stat(filename, &fs) != 0)
